@@ -17,11 +17,11 @@
   "types": "./dist/main.d.ts",
   "sideEffects": false,
   "devDependencies": {
-    "@vitest/coverage-v8": "^0.32.2",
-    "happy-dom": "^9.20.3",
-    "tsup": "^7.1.0",
-    "typescript": "^5.0.4",
-    "vitest": "^0.32.2"
+    "@vitest/coverage-v8": "^0.34.6",
+    "happy-dom": "^12.10.3",
+    "tsup": "^8.0.1",
+    "typescript": "^5.3.2",
+    "vitest": "^0.34.6"
   },
   "author": {
     "name": "Luca Barone",
