@@ -2,6 +2,7 @@
   "name": "@cloudwalker/react-inspect",
   "version": "0.9.0",
   "description": "Because json stringify is not enough",
+  "type": "module",
   "scripts": {
     "types": "tsc --noEmit",
     "lint": "eslint --ignore-path=\"../../.gitignore\" .",
@@ -42,26 +43,26 @@
   },
   "devDependencies": {
     "@cloudwalker/react-utils": "*",
-    "@storybook/addon-essentials": "^7.0.23",
-    "@storybook/addon-interactions": "^7.0.23",
-    "@storybook/addon-links": "^7.0.23",
-    "@storybook/react": "^7.0.23",
-    "@storybook/react-vite": "^7.0.23",
-    "@testing-library/react": "^14.0.0",
-    "@types/react": "^18.2.14",
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
-    "@vitejs/plugin-react": "^4.0.1",
-    "@vitest/coverage-v8": "^0.32.2",
-    "eslint": "^8.43.0",
-    "eslint-plugin-react": "^7.32.2",
+    "@storybook/addon-essentials": "^7.5.3",
+    "@storybook/addon-interactions": "^7.5.3",
+    "@storybook/addon-links": "^7.5.3",
+    "@storybook/react": "^7.5.3",
+    "@storybook/react-vite": "^7.5.3",
+    "@testing-library/react": "^14.1.2",
+    "@types/react": "^18.2.38",
+    "@typescript-eslint/eslint-plugin": "^6.12.0",
+    "@typescript-eslint/parser": "^6.12.0",
+    "@vitejs/plugin-react": "^4.2.0",
+    "@vitest/coverage-v8": "^0.34.6",
+    "eslint": "^8.54.0",
+    "eslint-plugin-react": "^7.33.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sanitize.css": "^13.0.0",
-    "storybook": "^7.0.23",
-    "tsup": "^7.1.0",
-    "typescript": "^5.0.4",
-    "vitest": "^0.32.2"
+    "storybook": "^7.5.3",
+    "tsup": "^8.0.1",
+    "typescript": "^5.3.2",
+    "vitest": "^0.34.6"
   },
   "dependencies": {
     "just-is-circular": "^2.1.1"
