@@ -32,9 +32,9 @@
     "react": ">=16.8"
   },
   "devDependencies": {
-    "@types/react": "^18.2.14",
+    "@types/react": "^18.2.38",
     "react": "^18.2.0",
-    "tsup": "^7.1.0",
-    "typescript": "^5.0.4"
+    "tsup": "^8.0.1",
+    "typescript": "^5.3.2"
   }
 }
