@@ -25,8 +25,8 @@
     "url": "https://github.com/cloud-walker"
   },
   "devDependencies": {
-    "tsup": "^7.1.0",
-    "typescript": "^5.0.4"
+    "tsup": "^8.0.1",
+    "typescript": "^5.3.2"
   },
   "publishConfig": {
     "access": "public"
