@@ -36,7 +36,7 @@ describe('DataHandler component', () => {
     expect(container).toMatchInlineSnapshot(`
       <div>
         <span
-          style="font-weight: bold;"
+          style="font-weight: bold; color: orange;"
         >
           666
         </span>
@@ -65,7 +65,7 @@ describe('DataHandler component', () => {
           style="cursor: pointer;"
         >
           <span
-            style="font-weight: bold;"
+            style="font-weight: bold; color: magenta;"
           >
             fn
           </span>
